Move the editor secion into more generic location.
